refactor(rust): move send-sync helpers to owners
